implemented adaptivity and Hot Rod, both for adaptive and fixed step sizes, but only time-serial into the non-MPI controller
